first commit o

+++ src/exercise_1/conditions/AmstrongNumWithLength_14_2.java
@@ -1,8 +1,10 @@
 package exercise_1.conditions;
 
 import java.util.Scanner;
+
+
 // JR: review done - Good
-//     TODO: 1) always check the warning given by IDE and try to fix it.
+//     TODO: 1) always check the warning given by IDE and try to fix it. - done
 public class AmstrongNumWithLength_14_2 {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
@@ -18,7 +20,7 @@ public class AmstrongNumWithLength_14_2 {
 
         while (temp > 0) {
             int a = temp % 10;
-            sum += Math.pow(a, length);
+            sum += (int) Math.pow(a, length);
             temp /= 10;
         }
